Cierre de turno con actualización de valor total. Cambio de tipo de campo a float en turno

+++ PuntoDeVenta.Domain/Entities/Turno.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public int CantidadInicio { get; set; }
         public int? CantidadFin { get; set; }
-        public int? ValorTotal { get; set; }
-        public int? ValorGanancia { get; set; }
+        public float? ValorTotal { get; set; }
+        public float? ValorGanancia { get; set; }
         public bool Finalizado { get; set; }
         public DateTime FechaInicio { get; set; }
         public DateTime? FechaFin { get; set; }
